add cert options
